Add detach and pid-file to restore
